requirements.txt
fastapi
uvicorn
jinja2
soundfile
numpy
scipy
requests
openai
python-multipart
python-dotenv
